@@ -1,6 +1,6 @@
 {
   "name": "choerodon-front-boot",
-  "version": "0.7.3",
+  "version": "0.8.0",
   "description": "Tools to build website for Choerodon",
   "main": "lib/containers/components/index.js",
   "homepage": "http://choerodon.io/",
@@ -55,12 +55,12 @@
     "case-sensitive-paths-webpack-plugin": "^2.1.2",
     "chalk": "^2.4.1",
     "change-case": "^3.0.1",
-    "choerodon-ui": "0.3.8",
+    "choerodon-ui": "0.4.0",
     "classnames": "^2.2.5",
     "commander": "^2.9.0",
     "component-classes": "^1.x",
     "css-loader": "^0.25.0",
-    "eslint": "^5.4.0",
+    "eslint": "^5.5.0",
     "eslint-config-airbnb": "latest",
     "eslint-plugin-babel": "^5.1.0",
     "eslint-plugin-import": "^2.14.0",
@@ -98,7 +98,7 @@
     "resolve": "^1.1.7",
     "rimraf": "2.4.3",
     "run-sequence": "^1.2.2",
-    "rxjs": "^5.5.6",
+    "rxjs": "^5.5.12",
     "sass-loader": "^6.0.6",
     "style-loader": "0.13.1",
     "stylelint": "^8.4.0",
