@@ -65,9 +65,14 @@ export default function getWebpackCommonConfig(mode, env) {
         minimize: true,
       }),
       new UglifyJsPlugin({
+        parallel: true,
+        cache: true,
         uglifyOptions: {
           output: {
-            ascii_only: true,
+            comments: false,
+          },
+          compress: {
+            warnings: false,
           },
         },
       }),
@@ -82,11 +87,12 @@ export default function getWebpackCommonConfig(mode, env) {
       modules: ['node_modules', join(__dirname, '../../node_modules')],
       extensions: ['.web.tsx', '.web.ts', '.web.jsx', '.web.js', '.ts', '.tsx', '.js', '.jsx', '.json'],
     },
-
     resolveLoader: {
       modules: ['node_modules', join(__dirname, '../../node_modules')],
     },
-
+    node: {
+      fs: 'empty',
+    },
     module: {
       noParse: [/moment.js/],
       rules: [
