@@ -3,7 +3,7 @@ import url from 'url';
 import { authorize, logout } from './authorize';
 import { getAccessToken, removeAccessToken, setAccessToken } from './accessToken';
 import { getCookie, removeCookie, setCookie } from './cookie';
-import { ACCESS_TOKEN, AUTH_URL, FILE_SERVER, LOCAL, USE_DASHBOARD } from './constants';
+import { ACCESS_TOKEN, AUTH_URL, FILE_SERVER, LOCAL, USE_DASHBOARD, WEBSOCKET_SERVER } from './constants';
 import { getMessage, intl } from './intl';
 import checkPassword from './checkPassword';
 import warning from '../../common/warning';
@@ -70,6 +70,8 @@ function fileServer(path) {
 export {
   ACCESS_TOKEN,
   AUTH_URL,
+  FILE_SERVER,
+  WEBSOCKET_SERVER,
   fileServer,
   setCookie,
   getCookie,
