@@ -2,8 +2,10 @@ import React, { Component } from 'react';
 import { Link, withRouter } from 'react-router-dom';
 import { inject, observer } from 'mobx-react';
 import { Button } from 'choerodon-ui';
+import { PREFIX_CLS } from '../../common/constants';
+import { dashboard } from '../../common';
 
-const prefixCls = 'header-logo';
+const prefixCls = `${PREFIX_CLS}-boot-header-logo`;
 
 @withRouter
 @inject('AppState', 'MenuStore')
@@ -15,9 +17,19 @@ export default class Logo extends Component {
   };
 
   render() {
-    const { MenuStore, location } = this.props;
+    const { AppState, MenuStore, location } = this.props;
     const { pathname, search } = location;
     const menus = MenuStore.getMenuData;
+    let homePath = '/';
+    if (dashboard) {
+      const { type, id, name, organizationId } = AppState.currentMenuType;
+      if (type && type !== 'site') {
+        homePath = `${homePath}?type=${type}&id=${id}&name=${name}`;
+        if (organizationId) {
+          homePath += `&organizationId=${organizationId}`;
+        }
+      }
+    }
     return (
       <div className={`${prefixCls}-wrap`}>
         {
@@ -28,7 +40,7 @@ export default class Logo extends Component {
         {
           pathname === '/' && !search
             ? <div className={prefixCls} />
-            : <Link to="/" className={prefixCls} />
+            : <Link to={homePath} className={prefixCls} />
         }
       </div>
     );
