@@ -6,12 +6,14 @@ import { createBrowserHistory } from 'history';
 import { configure } from 'mobx';
 import { observer, Provider } from 'mobx-react';
 import queryString from 'query-string';
-import stores from '../{{ source }}/containers/stores/index';
+import stores from '../{{ source }}/containers/stores';
 import AppState from '../{{ source }}/containers/stores/AppState';
 import asyncRouter from '../{{ source }}/containers/components/util/asyncRouter';
 import asyncLocaleProvider from '../{{ source }}/containers/components/util/asyncLocaleProvider';
-import { authorize, getAccessToken, setAccessToken } from '../{{ source }}/containers/common/index';
-import '../{{ source }}/containers/components/style/index';
+import { authorize, getAccessToken, setAccessToken, WEBSOCKET_SERVER } from '../{{ source }}/containers/common';
+import WSProvider from '../{{ source }}/containers/components/ws/WSProvider';
+import PermissionProvider from '../{{ source }}/containers/components/permission/PermissionProvider';
+import '../{{ source }}/containers/components/style';
 
 async function auth() {
   const { access_token: accessToken, token_type: tokenType, expires_in: expiresIn } = queryString.parse(window.location.hash);
@@ -42,13 +44,15 @@ class App extends Component {
       <UILocaleProviderAsync>
         <IntlProviderAsync>
           <Provider {...stores}>
-            <div>
-              <Router hashHistory={createBrowserHistory}>
-                <Switch>
-                  <Route path="/" component={Masters} />
-                </Switch>
-              </Router>
-            </div>
+            <PermissionProvider>
+              <WSProvider server={WEBSOCKET_SERVER}>
+                <Router hashHistory={createBrowserHistory}>
+                  <Switch>
+                    <Route path="/" component={Masters} />
+                  </Switch>
+                </Router>
+              </WSProvider>
+            </PermissionProvider>
           </Provider>
         </IntlProviderAsync>
       </UILocaleProviderAsync>
