@@ -1,4 +1,4 @@
-#!/usr/bin/python3
+#!/usr/bin/env python
 # _*_ coding:utf-8 _*_
 
 import sys
