@@ -1,4 +1,4 @@
-#!/usr/bin/python3
+#!/usr/bin/env python
 # _*_ coding:utf-8 _*_
 
 import json
@@ -78,12 +78,13 @@ def insertDb():
     contentConfig = yaml.load(ymlFile)
     insertDashboard(contentConfig)
     insertDashbaordTl(contentConfig)
+    insertDashboardRole(contentConfig)
     cursor.close()
     db.close()
     ymlFile.close()
 
 def returnId(table, code, namespace):
-    sql = "select ID from {table} where CODE='{code}' and NAMESPACE='{namespace}'".format(table=table, code=code, namespace=namespace)
+    sql = "SELECT ID FROM {table} WHERE CODE='{code}' AND NAMESPACE='{namespace}'".format(table=table, code=code, namespace=namespace)
     cursor.execute(sql)
     Id = cursor.fetchone()
     return Id
@@ -97,8 +98,8 @@ def insertDashboard(data):
             dashboard = dashboards[i]
             Id = returnId(table, dashboard["code"], dashboard["namespace"])
             if Id:
-                sql = "update {table} set CODE='{code}', FD_LEVEL='{level}', ICON='{icon}', SORT='{sort}', NAMESPACE='{namespace}'"
-                sql = (sql + " where CODE='{code}' and FD_LEVEL='{level}'").format(
+                sql = "UPDATE {table} SET CODE='{code}', FD_LEVEL='{level}', ICON='{icon}', SORT='{sort}', NAMESPACE='{namespace}'"
+                sql = (sql + " WHERE CODE='{code}' AND FD_LEVEL='{level}'").format(
                     table=table,
                     code=dashboard["code"],
                     namespace=dashboard["namespace"],
@@ -107,7 +108,7 @@ def insertDashboard(data):
                     sort=dashboard["sort"])
                 cursor.execute(sql)
             else:
-                sql = "insert into {table} (CODE, NAME, FD_LEVEL, TITLE, DESCRIPTION, ICON, NAMESPACE, SORT) values ('{code}', '{name}', '{level}', '{title}', '{description}', '{icon}', '{namespace}', '{sort}')"
+                sql = "INSERT INTO {table} (CODE, NAME, FD_LEVEL, TITLE, DESCRIPTION, ICON, NAMESPACE, SORT) VALUES ('{code}', '{name}', '{level}', '{title}', '{description}', '{icon}', '{namespace}', '{sort}')"
                 sql = sql.format(
                     table=table,
                     code=dashboard["code"],
@@ -131,7 +132,7 @@ def insertDashbaordTl(data):
             dashboard = dashboards[i]
             Id = returnId("IAM_DASHBOARD", dashboard["code"], dashboard["namespace"])
             if Id:
-                sql = "select ID from {table} where id={id}".format(
+                sql = "SELECT ID FROM {table} WHERE ID={id}".format(
                         table=table,
                         id=Id["ID"])
                 count = cursor.execute(sql)
@@ -143,16 +144,47 @@ def insertDashbaordTl(data):
                     updateTl(table, 'zh_CN', Id["ID"], dataLanguageChinese[i])
     except:
         dealFault()
+def insertDashboardRole(data):
+    try:
+        dashboards = data["dashboard"]
+        table = "IAM_DASHBOARD_ROLE"
+        for i in dashboards:
+            dashboard = dashboards[i]
+            Id = returnId("IAM_DASHBOARD", dashboard["code"], dashboard["namespace"])
+            if Id:
+                if "roles" in dashboard:
+                    roles = dashboard["roles"]
+                    for role in roles:
+                        sql = "SELECT ID FROM IAM_ROLE WHERE CODE='{code}' AND FD_LEVEL='{level}'".format(code=role, level=dashboard["level"]);
+                        cursor.execute(sql)
+                        roleId = cursor.fetchone()
+                        if roleId:
+                            sql = "SELECT ID FROM IAM_DASHBOARD_ROLE WHERE DASHBOARD_ID='{dashboardId}' AND ROLE_ID='{roleId}'".format(
+                                table=table,
+                                dashboardId=Id["ID"],
+                                roleId=roleId["ID"]
+                            )
+                            cursor.execute(sql)
+                            select = cursor.fetchone()   
+                            if not select:
+                                sql = "INSERT INTO {table} (DASHBOARD_ID, ROLE_ID) VALUES ('{dashboardId}', '{roleId}')".format(
+                                    table=table,
+                                    dashboardId=Id["ID"],
+                                    roleId=roleId["ID"]
+                                )
+                                cursor.execute(sql)
+    except:
+        dealFault()
 
 def insertTl(table, lang, id, name):
-    sql = "insert into {table} (LANG, ID, NAME) values ('{lang}','{id}','{name}')".format(
+    sql = "INSERT INTO {table} (LANG, ID, NAME) VALUES ('{lang}','{id}','{name}')".format(
         table=table,
         lang=lang,
         id=id,
         name=name)
     cursor.execute(sql)
 def updateTl(table, lang, id, name):
-    sql = "update {table} set ID='{id}', NAME='{name}' where ID={id} and LANG='{lang}'".format(
+    sql = "UPDATE {table} SET ID='{id}', NAME='{name}' WHERE ID={id} AND LANG='{lang}'".format(
         table=table,
         lang=lang,
         id=id,
